requirements.txt
python-dotenv
requests
bs4
transformers
torch
datasets
numpy
accelerate
gradio